+++ dat/outfits/bioship/skills/feral_rage_i.lua
@@ -50,7 +50,7 @@ function update( _p, po, dt )
    end
 end
 
-function onhit( p, po, armour, _shield )
+function onhit( p, po, armour )
    if not mem.active and armour > 0 then
       -- Already running, so just reset timer
       if mem.timer and mem.timer > 0 then