+++ dat/outfits/utility/cloaking_device.lua
@@ -97,7 +97,7 @@ function update( p, po, dt )
 end
 
 -- Disable on hit
-function onhit( p, po, _armour, _shield )
+function onhit( p, po )
    if mem.active then
       turnoff( p, po )
    end