+++ Changelog.md
@@ -11,7 +11,9 @@
    * Added weapon and launcher damage as disable ship stats.
    * Rewrote the following in rust
        - Lua Transform/Texture/Data/File modules
-   * Lua outfit "onanyimpact" reports armour, shield, and disable damage done
+   * Lua outfit scripting changes:
+       - "onanyimpact" reports armour, shield, and disable damage done
+       - "onhit" reports disable damage too [breaking change!!]
 
 
 ## v0.13.1 (unreleased)